Update and delete item functionality, refactored to separate action creators into separate files for clarity
