+++ src/App.js
@@ -1,15 +1,15 @@
 import React, { Component } from 'react';
 import { Route } from 'react-router-dom';
-import { Nav, Home, Authentication, AddInventory } from './components';
+import { Nav, Home, Authentication, AddItem } from './components';
 
 class App extends Component {
   render() {
     return (
       <>
         <Nav />
-        <Route exact path="/" component={props => <Home {...props} />} />
+        <Route exact path="/" render={props => <Home {...props} />} />
         <Route path="/auth" render={props => <Authentication {...props} />} />
-        <Route path="/add" component={AddInventory} />
+        <Route path="/add-item" component={AddItem} />
       </>
     );
   }
