--- src/components/AddInventory.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const AddInventory = props => <div>ADD INVENTORY</div>;
-
-export default AddInventory;
