+++ src/components/AddItem/AddItem.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const AddItem = props => <div>ADD INVENTORY</div>;
+
+export default AddItem;
