+++ src/components/Home/Inventory.js
