+++ src/components/index.js
@@ -1,5 +1,4 @@
 export { default as Nav } from './Nav';
-export { default as Home } from './Home';
-export { default as AddInventory } from './AddInventory';
-export { default as Register } from './Authentication';
+export { default as Home } from './Home/Home';
+export { default as AddItem } from './AddItem/AddItem';
 export { default as Authentication } from './Authentication';
