+++ src/redux/reducers/index.js
@@ -3,8 +3,10 @@ import {
   ERROR,
   REGISTRATION_SUCCESS,
   LOGIN_SUCCESS,
+  DELETE_USER_SUCCESS,
   SET_OPERATION,
   GET_INVENTORY_SUCCESS,
+  UPDATE_ITEM_SUCCESS,
   DELETE_ITEM_SUCCESS,
   ADD_ITEM_SUCCESS,
 } from '../actions';
@@ -49,7 +51,9 @@ const reducer = (state = initialState, action) => {
         ...state,
         loading: false,
         inventory: action.payload,
-        categories: Array.from(new Set(action.payload.map(x => x.category_id))),
+        categories: Array.from(
+          new Set(action.payload.map(item => item.category_id))
+        ),
         error: null,
       };
     case ADD_ITEM_SUCCESS:
@@ -57,15 +61,43 @@ const reducer = (state = initialState, action) => {
         ...state,
         loading: false,
         inventory: action.payload,
-        categories: Array.from(new Set(action.payload.map(x => x.category_id))),
+        categories: Array.from(
+          new Set(action.payload.map(item => item.category_id))
+        ),
         error: null,
       };
-    case DELETE_ITEM_SUCCESS:
+    case UPDATE_ITEM_SUCCESS: {
+      const newInventory = state.inventory.map(item => {
+        return item.id === action.payload.id ? action.payload : item;
+      });
       return {
         ...state,
         loading: false,
-        inventory: action.payload,
-        categories: Array.from(new Set(action.payload.map(x => x.category_id))),
+        inventory: newInventory,
+        categories: Array.from(
+          new Set(newInventory.map(item => item.category_id))
+        ),
+        error: null,
+      };
+    }
+    case DELETE_ITEM_SUCCESS: {
+      const newInventory = state.inventory.filter(
+        item => item.id !== action.payload
+      );
+      return {
+        ...state,
+        loading: false,
+        inventory: newInventory,
+        categories: Array.from(
+          new Set(newInventory.map(item => item.category_id))
+        ),
+        error: null,
+      };
+    }
+    case DELETE_USER_SUCCESS: // Necessary to update state?  How to indicate success?
+      return {
+        ...state,
+        loading: false,
         error: null,
       };
     default:
